pyproject.toml
[project]
name = "robot-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camera-sync",
    "duck-factory",
    "ikfast",
    "tornado>=6.4.2",
    "urbasic",
]

[tool.uv.sources]
duck-factory = { git = "https://github.com/6figuress/DuckFactory.git", rev = "develop" }
urbasic = { git = "https://github.com/6figuress/ur3e-control.git" }
ikfast = { git = "https://github.com/6figuress/ur3e-ik.git" }
camera-sync = { git = "https://github.com/6figuress/CameraCalibration.git" }

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
